requirements.txt
pymongo==4.10.1
pydantic==2.10.5
google-cloud-secret-manager==2.22.1
google-cloud-pubsub==2.28.0